feat(platform): decompose lesson component transformation
